js/xpconnect/tests/mochitest/file_nac.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
  <bindings xmlns="http://www.mozilla.org/xbl"
            xmlns:html="http://www.w3.org/1999/xhtml">
    <binding id="testBinding2">
      <implementation>
        <constructor>
          document.body.appendChild(parent.wrappedJSObject.nac);
          parent.wrappedJSObject.ok(true, "Didn't crash");
          parent.wrappedJSObject.SimpleTest.finish();
        </constructor>
      </implementation>
    </binding>
  </bindings>
<script type="application/javascript">
</script>
</head>
<body>
<div id="bindingSink" style="-moz-binding: url(#testBinding2);"></div>
</body>
</html>
